move env from file to builtins.c and rm env file
